chore: release v0.4.1 with security fixes and refactoring
- Security: Use UUID v4 for MCP request IDs
- Security: Add concurrency limiting to MCP client
- Refactor: Split client configuration into src/client/config.rs
- Refactor: Simplify retry logic and fix double-request bug
- Fix: Resolve clippy warnings in chat types
- Docs: Update README and CHANGELOG
